+++ src/main/java/bowlingexercise3/Ball.java
@@ -1,0 +1,6 @@
+package bowlingexercise3;
+
+public class Ball {
+
+
+}
